// src/types/car.ts
export interface Car {
  Make: string;
  Model: string;
  Year: string;
  Series: string;
  Price: string;
  'Other Price': string;
  'Photo Url List': string;
  VIN: string;
  'Dealer Name': string;
  'Dealer City': string;
  'Dealer Address': string;
  Status: string;
  Body: string;
  'Drivetrain Desc': string;
  Odometer: string;
  'New/Used': string;
  Vehicle: string;
  'Interior Color': string;
  Colour: string;
  'Fuel': string;
  'Transmission': string;
  'Engine': string;
  'Certified': string;
  'Stock #': string;
  'City MPG': string;
  'Highway MPG': string;
  Comments: string;
} 